// src/pages/IntelligentSearch/detail/data.d.ts
export interface ExampleListItem {
  id: number,
  name: String,
  type: String,
  object: String,
  part: String,
  category: String,
  procedure: String,
  example: String,
  time: number,
  source: String,
}

export interface MethodItem {
  id: number,
  methodName: String,
  icon: String,
  resource: String,
  uploadTime: String,
  methodContent: String,
  methodNumber: string,
  createTime: string,
  zjUse: string,
  methodCategory: string,
  methodType: string,
  processDesc: string,
  author: any,
  theory: string,
  demand: string,
  processChart: string,
  likeStatus: string,
  likeCount: string,
  storeStatus: string,
  storeCount: string,
  commentCount: string,
  showComment: boolean,
}